Fix optimizer JSON round-trip and exported-expression mangling
- as_long()/read_json_index() returned `long` and used std::stol, which
  is 32-bit on MSVC. Any 64-bit Index value in JSON (e.g. the default
  maximum_validation_failures = INT64_MAX) overflowed and threw
  "stol argument out of range", so optimizer JSON round-trip was broken
  on Windows. Widen to `long long` / std::stoll.
- Adam to/from_JSON now serialize learning_rate, beta_1, beta_2 and
  gradient_clip_norm (previously only BatchSize survived; a saved model
  silently reverted to defaults). Reads are guarded with has() so older
  JSON still loads.
- SGD to/from_JSON now serialize initial_learning_rate, initial_decay,
  momentum and nesterov instead of hard-coding momentum to 0.9/0.0 on
  load; ApplyMomentum is kept for backward compatibility.
- ModelExpression::replace_reserved_keywords used substring replacement
  for min/max/exp/tanh, mangling any variable name that contained them
  (maximum -> ma_ximum) and breaking exported code. Use whole-word
  replacement.

Add AdaptiveMomentEstimationTest.HyperparametersRoundTripJSON, which is
what surfaced the stol overflow.

Verified: build green; optimizer suites pass; full-suite diff vs a
stashed baseline shows no new non-GPU failures (the remaining failures
are the pre-existing CUDA-context and constructor-count failures).

+++ opennn/json.cpp
@@ -79,14 +79,14 @@ string Json::as_string() const
     return "";
 }
 
-long Json::as_long() const
+long long Json::as_long() const
 {
     using enum Kind;
     switch (kind)
     {
-    case Number: return long(number_value);
+    case Number: return (long long)(number_value);
     case Bool:   return bool_value ? 1 : 0;
-    case String: return string_value.empty() ? 0L : std::stol(string_value);
+    case String: return string_value.empty() ? 0LL : std::stoll(string_value);
     case Null:
     case Array:
     case Object: return 0;
@@ -531,7 +531,7 @@ float read_json_float(const Json* root, const string& field)
     return v ? float(v->as_double()) : 0.0f;
 }
 
-long read_json_index(const Json* root, const string& field)
+long long read_json_index(const Json* root, const string& field)
 {
     if (!root) return 0;
     const Json* v = root->find(field);
