+++ opennn/json.h
@@ -60,7 +60,7 @@ public:
     Json& set(const string& key, Json value);
     void push_back(Json value);
     string as_string() const;
-    long        as_long()   const;
+    long long   as_long()   const;
     double      as_double() const;
     bool        as_bool()   const;
     static Json  parse(const string& text);
@@ -105,7 +105,7 @@ void add_json_field(JsonWriter& writer,
 void write_json(JsonWriter& writer,
                 initializer_list<pair<const char*, string>> props);
 float       read_json_float   (const Json* root, const string& field);
-long        read_json_index  (const Json* root, const string& field);
+long long   read_json_index  (const Json* root, const string& field);
 bool        read_json_bool   (const Json* root, const string& field);
 string read_json_string (const Json* root, const string& field);
